+++ CHANGELOG.md
@@ -52,6 +52,9 @@ Bug Fixes
   not quite, flat (#741)
 - Fixed a bug in chromatic parametric galaxy models based on COSMOS galaxies.
   (#745)
+- Added the FourierSqrt operator to compute the Fourier-space square root of a
+  profile.  This is useful in implementing optimal coaddition algorithms; see
+  demo14.py (#748).
 
 Deprecated Features
 -------------------